Move log formatting to utils
